Create spawn-faerie.ps1
Create a vulnerable active directory system that can be used to test against most active directory attacks. Useful for those that want to build there AD-experience for the PWK/PEN-200.

Script was modified to build an AD box for the red team village CTF for Defcon 30! 

Currently tested and working on Windows Server 2016